# Zero-allocation isone